test: add functional tests

+++ routes/api.js
@@ -7,25 +7,20 @@
 */
 
 'use strict';
-
 const mongoose = require('mongoose');
-const Book = mongoose.model('Book', new mongoose.Schema({
-  title: { type: String, required: true },
-  comments: [String]
-}));
+const Book = require('../models/Book');
 
 module.exports = function (app) {
 
   app.route('/api/books')
     .get(async function (req, res) {
-      //response will be array of book objects
-      //json res format: [{"_id": bookid, "title": book_title, "commentcount": num_of_comments },...]
       try {
         const books = await Book.find();
+
         res.json(books.map(book => ({
-          _id: book._id,
+          _id: book._id.toString(),
           title: book.title,
-          commentcount: book.comments.length
+          comments: book.comments
         })));
       } catch (err) {
         res.status(500).send('Server error');
@@ -33,15 +28,18 @@ module.exports = function (app) {
     })
 
     .post(async function (req, res) {
-      let title = req.body.title;
-      //response will contain new book object including atleast _id and title
-      if (!title) {
+      if (!req.body || !('title' in req.body)) {
         return res.status(400).send('missing required field title');
       }
+
+      const title = req.body.title;
+      if (title === '') {
+        return res.status(400).send('missing required field title');
+      }
+
       try {
-        const newBook = new Book({ title, comments: [] });
-        await newBook.save();
-        res.json({ _id: newBook._id, title: newBook.title });
+        const newBook = await new Book({ title, comments: [] }).save();
+        res.json({ _id: newBook._id.toString(), title: newBook.title, comments: newBook.comments });
       } catch (err) {
         console.error('Error creating new book:', err);
         res.status(500).send('Server error');
@@ -49,10 +47,10 @@ module.exports = function (app) {
     })
 
     .delete(async function (req, res) {
-      //if successful response will be 'complete delete successful'
       try {
         await Book.deleteMany();
-        res.send('complete delete successful');
+
+        res.status(200).send('complete delete successful');
       } catch (err) {
         res.status(500).send('Server error');
       }
@@ -63,13 +61,18 @@ module.exports = function (app) {
   app.route('/api/books/:id')
     .get(async function (req, res) {
       const bookid = req.params.id;
-      //json res format: {"_id": bookid, "title": book_title, "comments": [comment,comment,...]}
+
+      if (!mongoose.Types.ObjectId.isValid(bookid)) {
+        return res.status(404).send('no book exists');
+      }
+
       try {
         const book = await Book.findById(bookid);
+
         if (!book) {
           return res.status(404).send('no book exists');
         }
-        res.json({ _id: book._id, title: book.title, comments: book.comments });
+        res.json({ _id: book._id.toString(), title: book.title, comments: book.comments });
       } catch (err) {
         res.status(500).send('Server error');
       }
@@ -78,18 +81,26 @@ module.exports = function (app) {
     .post(async function (req, res) {
       const bookid = req.params.id;
       const comment = req.body.comment;
-      //json res format same as .get
+
+      if (!mongoose.Types.ObjectId.isValid(bookid)) {
+        return res.status(404).send('no book exists');
+      }
+
       if (!comment) {
         return res.status(400).send('missing required field comment');
       }
+
       try {
         const book = await Book.findById(bookid);
+
         if (!book) {
           return res.status(404).send('no book exists');
         }
+
         book.comments.push(comment);
         await book.save();
-        res.json({ _id: book._id, title: book.title, comments: book.comments });
+
+        res.json({ _id: book._id.toString(), title: book.title, comments: book.comments });
       } catch (err) {
         res.status(500).send('Server error');
       }
@@ -97,12 +108,19 @@ module.exports = function (app) {
 
     .delete(async function (req, res) {
       const bookid = req.params.id;
+
+      if (!mongoose.Types.ObjectId.isValid(bookid)) {
+        return res.status(404).send('no book exists');
+      }
+
       try {
         const book = await Book.findByIdAndDelete(bookid);
+
         if (!book) {
           return res.status(404).send('no book exists');
         }
-        res.send('delete successful');
+
+        res.status(200).send('delete successful');
       } catch (err) {
         res.status(500).send('Server error');
       }
